FEAT: string pages

+++ common/hash.c
@@ -3,8 +3,8 @@
 #include <stdint.h>
 #include <string.h>
 
-size_t hash_fnv1a(const char* str, const size_t table_size) {
-	size_t hash_value = 2166136261u;
+uint32_t hash_fnv1a(const char* str, const uint32_t table_size) {
+	uint32_t hash_value = 2166136261u;
 	while (*str) {
 		hash_value ^= (unsigned char) *str++;
 		hash_value *= 16777619;
@@ -12,8 +12,8 @@ size_t hash_fnv1a(const char* str, const size_t table_size) {
 	return hash_value % table_size;
 }
 
-size_t hash_fnv1a_optimized(const char* str, const size_t table_size, const size_t sample_size) {
-	size_t hash_value = 2166136261u;
+uint32_t hash_fnv1a_optimized(const char* str, const uint32_t table_size, const uint32_t sample_size) {
+	uint32_t hash_value = 2166136261u;
 
 	// Hash first sample characters
 	for (size_t i = 0; i < sample_size && *str; i++) {
@@ -40,39 +40,39 @@ size_t hash_fnv1a_optimized(const char* str, const size_t table_size, const size
 	return hash_value % table_size;
 }
 
-size_t hash_djb2(const char* str, const size_t table_size) {
-	size_t hash_value = 5381;
+uint32_t hash_djb2(const char* str, const uint32_t table_size) {
+	uint32_t hash_value = 5381;
 	while (*str) {
 		hash_value = ((hash_value << 5) + hash_value) + *str++; // hash * 33 + char
 	}
 	return hash_value % table_size;
 }
 
-size_t hash_sdbm(const char* str, const size_t table_size) {
-	size_t hash_value = 0;
+uint32_t hash_sdbm(const char* str, const uint32_t table_size) {
+	uint32_t hash_value = 0;
 	while (*str) {
 		hash_value = *str++ + (hash_value << 6) + (hash_value << 16) - hash_value;
 	}
 	return hash_value % table_size;
 }
 
-size_t hash_murmur3_32(const char* str, const size_t table_size) {
-	size_t len = 0;
+uint32_t hash_murmur3_32(const char* str, const uint32_t table_size) {
+	uint32_t len = 0;
 	while (str[len] != '\0') {
 		len++;
 	}
 
-	const size_t seed = 2166136261u;
-	size_t h = seed;
-	const size_t c1 = 0xcc9e2d51;
-	const size_t c2 = 0x1b873593;
+	const uint32_t seed = 2166136261u;
+	uint32_t h = seed;
+	const uint32_t c1 = 0xcc9e2d51;
+	const uint32_t c2 = 0x1b873593;
 	const int r1 = 15;
 
-	const size_t nblocks = len / 4;
-	const size_t* blocks = (const size_t*)str;
+	const uint32_t nblocks = len / 4;
+	const uint32_t* blocks = (const uint32_t*)str;
 
-	for (size_t i = 0; i < nblocks; i++) {
-		size_t k = blocks[i];
+	for (uint32_t i = 0; i < nblocks; i++) {
+		uint32_t k = blocks[i];
 		k *= c1;
 		k = (k << r1) | (k >> (32 - r1));
 		k *= c2;
@@ -83,7 +83,7 @@ size_t hash_murmur3_32(const char* str, const size_t table_size) {
 	}
 
 	const uint8_t* tail = (const uint8_t*)(str + nblocks * 4);
-	size_t k1 = 0;
+	uint32_t k1 = 0;
 
 	switch (len & 3) {
 		case 3:
