+++ common/hash.h
@@ -2,15 +2,16 @@
 #define HASH_H
 
 #include <stddef.h>
+#include <stdint.h>
 
-size_t hash_fnv1a(const char* str, size_t table_size);
+uint32_t hash_fnv1a(const char* str, uint32_t table_size);
 
-size_t hash_fnv1a_optimized(const char* str, size_t table_size, size_t sample_size);
+uint32_t hash_fnv1a_optimized(const char* str, uint32_t table_size, uint32_t sample_size);
 
-size_t hash_djb2(const char* str, size_t table_size);
+uint32_t hash_djb2(const char* str, uint32_t table_size);
 
-size_t hash_sdbm(const char* str, size_t table_size);
+uint32_t hash_sdbm(const char* str, uint32_t table_size);
 
-size_t hash_murmur3_32(const char* str, size_t table_size);
+uint32_t hash_murmur3_32(const char* str, uint32_t table_size);
 
 #endif //HASH_H
